Added Error Handling

+++ controllers/post.controller.js
@@ -58,44 +58,52 @@ const getPostById = async(req, res, next) => {
 const addReactionToPost = async(req, res, next) => {
     try 
     {
-        const post = await Post.findOne({_id: req.body._id});
-        const newReaction = new React({
-            name: req.params.username,
-            type: req.params.react
-        });
-        let id = 0;
-        for (let reaction of post.reacts) 
-        {
-            if(reaction.name === newReaction.name) 
-            {
-                if(newReaction.type === reaction.type) 
-                {
-                    post[reaction.type]--;
-                }
-                else 
-                {
-                    post[reaction.type]--;
-                    post[newReaction.type]++;
-                    post.reacts.push(newReaction);
-                }
-                break;
-            }
-            id++;
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
         }
-        if(id === post.reacts.length) {
-            post[newReaction.type]++;
-            post.reacts.push(newReaction);
+        else if(req.user.username !== req.params.username) {
+            res.status(401).json({Error: "You are not authenticated"});
         }
         else {
-            post.reacts.splice(id, 1);
+            const post = await Post.findOne({_id: req.body._id});
+            const newReaction = new React({
+                name: req.params.username,
+                type: req.params.react
+            });
+            let id = 0;
+            for (let reaction of post.reacts) 
+            {
+                if(reaction.name === newReaction.name) 
+                {
+                    if(newReaction.type === reaction.type) 
+                    {
+                        post[reaction.type]--;
+                    }
+                    else 
+                    {
+                        post[reaction.type]--;
+                        post[newReaction.type]++;
+                        post.reacts.push(newReaction);
+                    }
+                    break;
+                }
+                id++;
+            }
+            if(id === post.reacts.length) {
+                post[newReaction.type]++;
+                post.reacts.push(newReaction);
+            }
+            else {
+                post.reacts.splice(id, 1);
+            }
+            post.save()
+            .then((data) => {
+                res.json(data);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
         }
-        post.save()
-        .then((data) => {
-            res.json(data);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
     }
     catch(error) {
         res.json(next(error));
@@ -104,27 +112,35 @@ const addReactionToPost = async(req, res, next) => {
 
 const editPost = async(req, res, next) => {
     try {
-        const post = await Post.findOne({_id: req.params.id});
-        var imageUrl = "";
-        if(req.body.data !== "") {
-            var fileStr = req.body.data;
-            var uploadedResponse = await cloudinary.uploader.
-            upload(fileStr, {
-                upload_preset: "socialites"
-            });
-            imageUrl = uploadedResponse.url;
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
         }
-        post.title = req.body.title;
-        post.content = req.body.content;
-        post.category = req.body.category;
-        post.imageUrl = imageUrl;
-        post.save()
-        .then((data) => {
-            res.json(data);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
+        else if(req.user.username !== req.body.author) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            const post = await Post.findOne({_id: req.params.id});
+            var imageUrl = "";
+            if(req.body.data !== "") {
+                var fileStr = req.body.data;
+                var uploadedResponse = await cloudinary.uploader.
+                upload(fileStr, {
+                    upload_preset: "socialites"
+                });
+                imageUrl = uploadedResponse.url;
+            }
+            post.title = req.body.title;
+            post.content = req.body.content;
+            post.category = req.body.category;
+            post.imageUrl = imageUrl;
+            post.save()
+            .then((data) => {
+                res.json(data);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
+        }
     }
     catch(error) {
         res.json(next(error));
@@ -133,35 +149,43 @@ const editPost = async(req, res, next) => {
 
 const addPost = async(req, res, next) => {
     try {
-        var imageUrl = "";
-        if(req.body.data !== "") {
-            var fileStr = req.body.data;
-            var uploadedResponse = await cloudinary.uploader.
-            upload(fileStr, {
-                upload_preset: "socialites"
-            });
-            imageUrl = uploadedResponse.url;
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
         }
-        const { author, title, content, category } = req.body;
-        const post = new Post({
-            author,
-            title,
-            content,
-            category,
-            imageUrl,
-            reacts: [],
-            comment_count: 0,
-            like: 0,
-            love: 0,
-            laugh: 0
-        });
-        post.save()
-        .then((data) => {
-            res.json(data);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
+        else if(req.user.username !== req.body.author) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            var imageUrl = "";
+            if(req.body.data !== "") {
+                var fileStr = req.body.data;
+                var uploadedResponse = await cloudinary.uploader.
+                upload(fileStr, {
+                    upload_preset: "socialites"
+                });
+                imageUrl = uploadedResponse.url;
+            }
+            const { author, title, content, category } = req.body;
+            const post = new Post({
+                author,
+                title,
+                content,
+                category,
+                imageUrl,
+                reacts: [],
+                comment_count: 0,
+                like: 0,
+                love: 0,
+                laugh: 0
+            });
+            post.save()
+            .then((data) => {
+                res.json(data);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
+        }
     }   
     catch(error) {
         res.json(next(error));
@@ -170,17 +194,25 @@ const addPost = async(req, res, next) => {
 
 const addComment = async(req, res, next) => {
     try {
-        const post = await Post.findOne({_id: req.params.id});
-        const comment = new Comment(req.body);
-        post.comments.push(comment);
-        post.comment_count = post.comments.length;
-        post.save()
-        .then((data) => {
-            res.json(data.comments);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else if(req.user.username !== req.body.name) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            const post = await Post.findOne({_id: req.params.id});
+            const comment = new Comment(req.body);
+            post.comments.push(comment);
+            post.comment_count = post.comments.length;
+            post.save()
+            .then((data) => {
+                res.json(data.comments);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
+        }
     }
     catch(error) {
         res.json(next(error));
@@ -189,47 +221,55 @@ const addComment = async(req, res, next) => {
 
 const addReactionToComment = async(req, res, next) => {
     try {
-        const post = await Post.findOne({_id: req.params.postId});
-        let index = await post.comments.findIndex((comment) => (comment._id == req.body._id));
-        const newReaction = new React({
-            name: req.params.username,
-            type: req.params.react
-        });
-        let id = 0;
-        for (let reaction of post.comments[index].reacts) 
-        {
-            if(reaction.name === newReaction.name)
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else if(req.user.username !== req.params.username) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            const post = await Post.findOne({_id: req.params.postId});
+            let index = await post.comments.findIndex((comment) => (comment._id == req.body._id));
+            const newReaction = new React({
+                name: req.params.username,
+                type: req.params.react
+            });
+            let id = 0;
+            for (let reaction of post.comments[index].reacts) 
             {
-                if(newReaction.type !== reaction.type)
+                if(reaction.name === newReaction.name)
                 {
-                    post.comments[index][reaction.type]--;
-                    post.comments[index][newReaction.type]++;
-                    post.comments[index].reacts.push(newReaction);
+                    if(newReaction.type !== reaction.type)
+                    {
+                        post.comments[index][reaction.type]--;
+                        post.comments[index][newReaction.type]++;
+                        post.comments[index].reacts.push(newReaction);
+                    }
+                    else 
+                    {
+                        post.comments[index][newReaction.type]--;
+                    }
+                    break;
                 }
-                else 
-                {
-                    post.comments[index][newReaction.type]--;
-                }
-                break;
+                id++;
             }
-            id++;
+            if(id === post.comments[index].reacts.length) 
+            {
+                await post.comments[index].reacts.push(newReaction);
+                post.comments[index][newReaction.type]++;
+            }
+            else 
+            {
+                await post.comments[index].reacts.splice(id,1);
+            }
+            post.save()
+            .then((data) => {
+                res.json(data.comments[index]);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
         }
-        if(id === post.comments[index].reacts.length) 
-        {
-            await post.comments[index].reacts.push(newReaction);
-            post.comments[index][newReaction.type]++;
-        }
-        else 
-        {
-            await post.comments[index].reacts.splice(id,1);
-        }
-        post.save()
-        .then((data) => {
-            res.json(data.comments[index]);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
     }
     catch(error) {
         res.json(next(error));
@@ -238,20 +278,28 @@ const addReactionToComment = async(req, res, next) => {
 
 const deleteComment = async(req, res, next) => {
     try {
-        const post = await Post.findOne({_id: req.params.id});
-        var index = post.comments.findIndex((comment) => comment._id == req.body._id);
-        if(index !== -1) 
-        {
-            post.comments.splice(index, 1);
-            post.comment_count--;
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
         }
-        post.save()
-        .then((data) => {
-            res.json(data.comments);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
+        else if(req.user.username !== req.params.username) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            const post = await Post.findOne({_id: req.params.id});
+            var index = post.comments.findIndex((comment) => comment._id == req.body._id);
+            if(index !== -1) 
+            {
+                post.comments.splice(index, 1);
+                post.comment_count--;
+            }
+            post.save()
+            .then((data) => {
+                res.json(data.comments);
+            })
+            .catch((err) => {
+                res.json(err);
+            });
+        }
     }
     catch(error) {
         res.json(next(error));
@@ -260,8 +308,16 @@ const deleteComment = async(req, res, next) => {
 
 const deletePost = async(req, res, next) => {
     try {
-        const response = await Post.deleteOne({_id: req.params.id});
-        res.json(response);
+        if(req.user === null) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else if(req.user.username !== req.params.username) {
+            res.status(401).json({Error: "You are not authenticated"});
+        }
+        else {
+            const response = await Post.deleteOne({_id: req.params.id});
+            res.json(response);
+        }
     }
     catch(error) {
         res.json(next(error));
